Update to reduce dependency on config files
CompileClassifierConfig now has a second constructor that takes a
network (.net or .uai) file (and other arguments), and pulls most of the relevant fields from
it.
RunCompiler has been updated to make use of this, and provide a command
line interface that uses network files instead of config files
CompileClassifier has been slightly updated, as the removed
functionality is now being done in CompileClassifierConfig
run has been updated with additional examples of how to run with the new
RunCompiler
ConfigGenerator has been created to generate config files for all the
roots of a particular network file
all_configs has been updated by a single character in order for it to be
a valid json file.
README has been updated to provide more usage detail. Most of the
comments are in the code, however.